+++ app/api/loans/[id]/route.ts
@@ -1,15 +1,22 @@
 // app/api/loans/[id]/route.ts
 
+import { auth } from '@/lib/auth';
 import { prisma } from '@/lib/prisma';
+import { headers } from 'next/headers';
 import { NextResponse } from 'next/server';
 
 export async function DELETE(
-  req: Request,
+  _req: Request,
   { params }: { params: Promise<{ id: string }> }
 ) {
-  console.log('DELETE ICI');
+  const session = await auth.api.getSession({
+    headers: await headers(),
+  });
+
+  if (!session?.user) {
+    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
+  }
   const { id } = await params;
-  console.log('id = ', id);
   try {
     const deletedLoan = await prisma.loan.delete({
       where: { id },
